+++ sqllin-driver/src/appleTest/kotlin/com/ctrip/sqllin/driver/PlatformApple.kt
@@ -20,7 +20,6 @@ import kotlinx.cinterop.UnsafeNumber
 import platform.Foundation.NSDocumentDirectory
 import platform.Foundation.NSSearchPathForDirectoriesInDomains
 import platform.Foundation.NSUserDomainMask
-import platform.posix.remove
 
 /**
  * Apple platform-related functions
@@ -30,5 +29,3 @@ import platform.posix.remove
 @OptIn(UnsafeNumber::class)
 actual fun getPlatformStringPath(): String =
     (NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, true).firstOrNull() as? String ?: "")
-
-actual fun deleteFile(file: String): Boolean = remove(file) == 0