+++ sqllin-driver/src/linuxTest/kotlin/com/ctrip/sqllin/driver/PlatformLinux.kt
@@ -18,7 +18,6 @@ package com.ctrip.sqllin.driver
 
 import kotlinx.cinterop.toKString
 import platform.posix.getcwd
-import platform.posix.remove
 
 /**
  * Linux platform-related functions
@@ -27,5 +26,3 @@ import platform.posix.remove
 
 actual fun getPlatformStringPath(): String =
     getcwd(null, 0)?.toKString() ?: throw IllegalStateException("The temp path created error")
-
-actual fun deleteFile(file: String): Boolean = remove(file) == 0