+++ sqllin-driver/src/mingwMain/kotlin/com/ctrip/sqllin/driver/platform/UtilsMinGW.kt
@@ -5,7 +5,7 @@ import kotlinx.cinterop.CPointer
 import kotlinx.cinterop.toKStringFromUtf8
 
 /**
- * The tools about bytes, implementation of MinGW
+ * The tools with Windows implementation
  * @author yqiao
  */