+++ sqllin-driver/src/nativeTest/kotlin/com/ctrip/sqllin/driver/NativeTest.kt
@@ -16,7 +16,6 @@
 
 package com.ctrip.sqllin.driver
 
-import com.ctrip.sqllin.driver.platform.separatorChar
 import kotlin.test.AfterTest
 import kotlin.test.Test
 
@@ -27,8 +26,8 @@ import kotlin.test.Test
 
 class NativeTest {
 
-    private val path = getPlatformStringPath()
-    private val commonTest = CommonBasicTest(path.toDatabasePath())
+    private val path = getPlatformStringPath().toDatabasePath()
+    private val commonTest = CommonBasicTest(path)
 
     @Test
     fun testCreateAndUpgrade() = commonTest.testCreateAndUpgrade()
@@ -50,13 +49,6 @@ class NativeTest {
 
     @AfterTest
     fun setDown() {
-        listOf(
-            "$path$separatorChar${SQL.DATABASE_NAME}",
-            "$path$separatorChar${SQL.DATABASE_NAME}-shm",
-            "$path$separatorChar${SQL.DATABASE_NAME}-wal",
-        ).forEach {
-            val result = deleteFile(it)
-            println("Delete file: $it, result: $result")
-        }
+        deleteDatabase(path, SQL.DATABASE_NAME)
     }
 }